# doc/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "2.0.0",
    "title": "ONG API",
    "description": "API desenvolvida para prover dados ao front-end do sistema de cadastros de uma ONG",
    "contact": {
      "email": "storeexodia@gmail.com"
    }
  },
  "host": "localhost:3333",
  "basePath": "/ong",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/cadastrarAcolhido": {
      "post": {
        "tags": [
          "Acolhidos"
        ],
        "summary": "Cadastra um novo acolhido no banco de dados.",
        "description": "Essa rota é responsável por cadastrar um novo acolhido.",
        "parameters": [
          {
            "name": "dados do acolhido",
            "in": "body",
            "description": "Informações do acolhido.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CadastroDoAcolhido"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/todosOsAcolhidos": {
      "get": {
        "tags": [
          "Acolhidos"
        ],
        "summary": "Retorna todos os acolhidos cadastrados",
        "description": "Retorna todos os acolhidos cadastrados",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/acolhidoPorMatricula": {
      "get": {
        "tags": [
          "Acolhidos"
        ],
        "summary": "Retorna um acolhido utilizando sua matrícula.",
        "description": "Retorna o acolhido que estiver a matrícula cadastrada no banco.",
        "parameters": [
          {
            "name": "matricula",
            "description": "Matrícula do acolhido.",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/acolhidosPorUnidade": {
      "get": {
        "tags": [
          "Acolhidos"
        ],
        "summary": "Retorna um acolhido utilizando a unidade que está alocado.",
        "description": "Retorna o acolhido que estiver na unidade informada.",
        "parameters": [
          {
            "name": "unidadeDeOrigem",
            "description": "Unidade que o acolhido está alocado.",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/deletarAcolhido": {
      "delete": {
        "tags": [
          "Acolhidos"
        ],
        "summary": "Remove um acolhido do banco de dados.",
        "description": "Remove um acolhido do banco de dados localizando-o através de sua matrícula.",
        "parameters": [
          {
            "name": "matricula",
            "description": "Matrícula do acolhido.",
            "required": true,
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/atualizarAcolhido/{matricula}": {
      "put": {
        "tags": [
          "Acolhidos"
        ],
        "summary": "Atualiza informações de um acolhido.",
        "description": "Atualiza as informações do acolhido, localizando-o através de sua matrícula, repassando os dados necessários. ",
        "parameters": [
          {
            "name": "matricula",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Matrícula do acolhido."
          },
          {
            "name": "dados atualizados",
            "in": "body",
            "description": "Acolhido atualizado.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CadastroDoAcolhido"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/cadastrarAcesso": {
      "post": {
        "tags": [
          "Credencial"
        ],
        "summary": "Cadastra uma nova credencial no banco de dados.",
        "description": "Essa rota é responsável por cadastrar uma nova credencial.",
        "parameters": [
          {
            "name": "dados da credencial",
            "in": "body",
            "description": "Informações da credencial.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CadastroDaCredencial"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/validarAcesso": {
      "post": {
        "tags": [
          "Credencial"
        ],
        "summary": "Validar credencial de acesso.",
        "description": "Validar cradencial de acesso do usuário. Retorna true caso a senha passada esteja correta e false para incorreta.",
        "parameters": [
          {
            "name": "Login e Senha passado pelo usuário",
            "in": "body",
            "description": "Senha passada pelo usuário do sistema.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CredencialPassada"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/buscarCadastro": {
      "get": {
        "tags": [
          "Credencial"
        ],
        "summary": "Retorna uma credencial pelo login.",
        "description": "Retorna uma credencial pelo login.",
        "parameters": [
          {
            "name": "login",
            "description": "Login da credencial.",
            "required": true,
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/buscarTodosOsCadastros": {
      "get": {
        "tags": [
          "Credencial"
        ],
        "summary": "Retorna todas as credenciais cadastradas",
        "description": "Retorna todas as credenciais cadastradas",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/deletarCadastro": {
      "delete": {
        "tags": [
          "Credencial"
        ],
        "summary": "Remove uma credencial do banco de dados.",
        "description": "Remove uma credencial do banco de dados localizando-o através de seu login.",
        "parameters": [
          {
            "name": "login",
            "description": "Login da credencial.",
            "required": true,
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/atualizarSenha/{login}": {
      "put": {
        "tags": [
          "Credencial"
        ],
        "summary": "Atualiza informações do usuário.",
        "description": "Atualiza as informações do usuário, localizando-o através de seu login, atualizando a senha. ",
        "parameters": [
          {
            "name": "login",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Login do usuário."
          },
          {
            "name": "dados atualizados",
            "in": "body",
            "description": "Senha atualizada",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CadastroDaCredencial"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/cadastrarDadosMedicos": {
      "post": {
        "tags": [
          "Medico"
        ],
        "summary": "Cadastra os dados médicos do acolhido.",
        "description": "Cadastra os dados médicos do acolhido.",
        "parameters": [
          {
            "name": "dados do acolhido",
            "in": "body",
            "description": "Informações do acolhido.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/DadosMedicos"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/getTodosDadosMedicos": {
      "get": {
        "tags": [
          "Medico"
        ],
        "summary": "Retorna os dados médicos de todos os acolhidos.",
        "description": "Retorna os dados médicos de todos os acolhidos.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/getUmDadoMedico": {
      "get": {
        "tags": [
          "Medico"
        ],
        "summary": "Retorna os dados médicos de um acolhido, utilizando o nome e a data de nascimento para localiza-lo.",
        "description": "Retorna os dados médicos de um acolhido, utilizando o nome e a data de nascimento para localiza-lo.",
        "parameters": [
          {
            "name": "nome",
            "description": "Nome do acolhido.",
            "required": true,
            "in": "query",
            "type": "string"
          },
          {
            "name": "dataNascimento",
            "description": "Data de nascimento do acolhido.",
            "required": true,
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/deletarDadosMedicos": {
      "delete": {
        "tags": [
          "Medico"
        ],
        "summary": "Remove os dados médicos de um acolhido.",
        "description": "Remove os dados médicos de um acolhido, localizando ele pelo nome e data de nascimento. ",
        "parameters": [
          {
            "name": "nome",
            "description": "Nome do acolhido.",
            "required": true,
            "in": "query",
            "type": "string"
          },
          {
            "name": "dataNascimento",
            "description": "Data de nascimento do acolhido.",
            "required": true,
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/atualizarDadosMedicos/": {
      "put": {
        "tags": [
          "Medico"
        ],
        "summary": "Atualiza informações dos dados médicos de um acolhido.",
        "description": "Atualiza as informações dos dados médicos de de um acolhido, localizando-o através de seu nome e data de nascimento, repassando os dados necessários. ",
        "parameters": [
          {
            "name": "nome",
            "description": "Nome do acolhido.",
            "required": true,
            "in": "query",
            "type": "string"
          },
          {
            "name": "dataNascimento",
            "description": "Data de nascimento do acolhido.",
            "required": true,
            "in": "query",
            "type": "string"
          },
          {
            "name": "dados atualizados",
            "in": "body",
            "description": "Dados médicos do acolhido atualizados.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/DadosMedicos"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "definitions": {
    "CadastroDoAcolhido": {
      "type": "object",
      "properties": {
        "nomeCompleto": {
          "type": "string",
          "example": "Quentin Tarantino"
        },
        "nomeDaMae": {
          "type": "string",
          "example": "Patrícia Amorim"
        },
        "nomeDoPai": {
          "type": "string",
          "example": "João Planaltina"
        },
        "naturalidade": {
          "type": "string",
          "example": "Brasileiro"
        },
        "nacionalidade": {
          "type": "string",
          "example": "Brasileiro"
        },
        "idade": {
          "type": "number",
          "example": 59
        },
        "dataDeNascimento": {
          "type": "string",
          "example": "05-05-2022"
        },
        "estadoCivil": {
          "type": "string",
          "example": "Solteiro"
        },
        "raca": {
          "type": "string",
          "example": "Branco"
        },
        "escolaridade": {
          "type": "string",
          "example": "Superior Completo"
        },
        "profissao": {
          "type": "string",
          "example": "Diretor de Cinema"
        },
        "CPF": {
          "type": "string",
          "example": "123.456.789-10"
        },
        "sexo": {
          "type": "string",
          "example": "Masculino"
        },
        "RG": {
          "type": "string",
          "example": "12.345.678"
        },
        "estadoEmissor": {
          "type": "string",
          "example": "PE"
        },
        "dataDeEmissaoRG": {
          "type": "string",
          "example": "05-05-2022"
        },
        "orgaoEmissor": {
          "type": "string",
          "example": "SDS"
        },
        "CEP": {
          "type": "string",
          "example": "52091-240"
        },
        "endereco": {
          "type": "string",
          "example": "Rua Córrego do boleiro, 221"
        },
        "estado": {
          "type": "string",
          "example": "PE"
        },
        "municipio": {
          "type": "string",
          "example": "Recife"
        },
        "bairro": {
          "type": "string",
          "example": "Nova Descoberta"
        },
        "unidadeDeOrigem": {
          "type": "string",
          "example": "Recife Centro"
        },
        "numeroDeInternacoes": {
          "type": "number",
          "example": 0
        },
        "contatosDeEmergencia": {
          "type": "object",
          "properties": {
            "numero1": {
              "type": "string",
              "example": "81 98342-6557"
            },
            "numero2": {
              "type": "string",
              "example": "81 98726-7365"
            }
          }
        },
        "dataDeSaida": {
          "type": "string",
          "example": "06-05-2022"
        },
        "matricula": {
          "type": "string",
          "example": "02307485"
        },
        "possuiFilhos": {
          "type": "boolean",
          "example": false
        },
        "situacaoDeRua": {
          "type": "boolean",
          "example": false
        },
        "tempoDeRua": {
          "type": "string",
          "example": "1 ano"
        },
        "tempoDeDrogadicao": {
          "type": "string",
          "example": "1 ano"
        },
        "tiposDeDrogas": {
          "type": "string",
          "example": "Maconha"
        },
        "usaMedicamentos": {
          "type": "boolean",
          "example": false
        },
        "doencasPrecisaCuidados": {
          "type": "boolean",
          "example": false
        },
        "alergia": {
          "type": "boolean",
          "example": false
        },
        "acompanhamentoPsiquiatrico": {
          "type": "boolean",
          "example": false
        },
        "respondeProcesso": {
          "type": "boolean",
          "example": false
        },
        "recebeBeneficio": {
          "type": "boolean",
          "example": false
        },
        "religiao": {
          "type": "string",
          "example": "Protestantismo"
        },
        "igrejaEvangelica": {
          "type": "boolean",
          "example": false
        },
        "batizado": {
          "type": "boolean",
          "example": false
        },
        "casaDeRecuperacao": {
          "type": "boolean",
          "example": false
        },
        "jaFoiAcolhido": {
          "type": "boolean",
          "example": false
        },
        "deficiencia": {
          "type": "boolean",
          "example": false
        },
        "antecedentesCriminais": {
          "type": "boolean",
          "example": false
        },
        "fazNasHorasVagas": {
          "type": "string",
          "example": ""
        },
        "consideracoesDeAtendimento": {
          "type": "string",
          "example": ""
        }
      }
    },
    "CadastroDaCredencial": {
      "type": "object",
      "properties": {
        "nome": {
          "type": "string",
          "example": "Valentina"
        },
        "sobrenome": {
          "type": "string",
          "example": "Silva"
        },
        "login": {
          "type": "string",
          "example": "01304487"
        },
        "counter": {
          "type": "string",
          "example": "123456"
        },
        "usuario": {
          "type": "string",
          "example": "voluntário"
        }
      }
    },
    "CredencialPassada": {
      "type": "object",
      "properties": {
        "login": {
          "type": "string",
          "example": "01304487"
        },
        "counter": {
          "type": "string",
          "example": "123456"
        }
      }
    },
    "DadosMedicos": {
      "type": "object",
      "properties": {
        "nome": {
          "type": "string",
          "example": "João"
        },
        "dataNascimento": {
          "type": "string",
          "example": "06-05-2022"
        },
        "dataChegada": {
          "type": "string",
          "example": "06-05-2022"
        },
        "dataAtendimento": {
          "type": "string",
          "example": "06-05-2022"
        },
        "aspectosCognitivos": {
          "type": "string",
          "example": "Info."
        },
        "opniaoSituacao": {
          "type": "string",
          "example": "Info."
        },
        "consideraDependenteQuimico": {
          "type": "string",
          "example": "Info."
        },
        "composicaoFamiliar": {
          "type": "string",
          "example": "Info."
        },
        "relacoesParentais": {
          "type": "string",
          "example": "Info."
        },
        "relacionamentoComFamilia": {
          "type": "string",
          "example": "Info."
        },
        "reestabelecimentoFamiliar": {
          "type": "string",
          "example": "Info."
        },
        "comoComecouNasDrogas": {
          "type": "string",
          "example": "Info."
        },
        "informacoesAdicionais": {
          "type": "string",
          "example": "Info."
        },
        "orientacoesPassadas": {
          "type": "boolean",
          "example": false
        },
        "contatosDeEmergencia": {
          "type": "object",
          "properties": {
            "numero1": {
              "type": "string",
              "example": "11111111"
            },
            "numero2": {
              "type": "string",
              "example": "11111111"
            }
          }
        },
        "evolucaoDoVinculo": {
          "type": "string",
          "example": "Info."
        },
        "reitegracaoFamiliar": {
          "type": "string",
          "example": "Info."
        },
        "antecedentesDependenciaQuimica": {
          "type": "boolean",
          "example": false
        },
        "necessitaApoioJuridico": {
          "type": "object",
          "properties": {
            "necessita": {
              "type": "boolean",
              "example": false
            },
            "especifique": {
              "type": "string",
              "example": ""
            }
          }
        },
        "rendaPropria": {
          "type": "object",
          "properties": {
            "possui": {
              "type": "boolean",
              "example": false
            },
            "queTipo": {
              "type": "string",
              "example": ""
            }
          }
        },
        "expProfissional": {
          "type": "string",
          "example": "Info."
        },
        "propostaDeRenda": {
          "type": "string",
          "example": "Info."
        },
        "propostaDeMoradia": {
          "type": "string",
          "example": "Info."
        },
        "objetivosDoAcolhido": {
          "type": "string",
          "example": "Info."
        },
        "tipoDeSaida": {
          "type": "string",
          "example": "Info."
        }
      }
    }
  }
}